Style: schedule 예외 처리

+++ src/main/java/com/example/springjpa/schedule/api/ScheduleController.java
@@ -1,5 +1,7 @@
 package com.example.springjpa.schedule.api;
 
+import com.example.springjpa.global.exception.custom.InvalidCredentialsException;
+import com.example.springjpa.global.jwt.TokenProvider;
 import com.example.springjpa.schedule.api.dto.ScheduleUpdateRequest;
 import com.example.springjpa.schedule.api.dto.request.AssignUsersRequest;
 import com.example.springjpa.schedule.api.dto.request.ScheduleSaveRequest;
@@ -7,7 +9,6 @@ import com.example.springjpa.schedule.api.dto.response.ScheduleResponse;
 import com.example.springjpa.schedule.application.ScheduleService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -15,27 +16,30 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.naming.AuthenticationException;
 import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/schedule")
 public class ScheduleController {
+    public static final String AUTHORIZATION_HEADER = "Authorization";
     private final ScheduleService scheduleService;
+    private final TokenProvider tokenProvider;
 
     @PostMapping
-    public ResponseEntity<ScheduleResponse> createSchedule(@RequestBody @Valid ScheduleSaveRequest request) {
-        // 추후 토큰받아서 작성자 정보 넘기기 - @AuthenticationPrincipal UserDetail userDetail
-        ScheduleResponse response = scheduleService.create(request);
+    public ResponseEntity<ScheduleResponse> createSchedule(@CookieValue(value = AUTHORIZATION_HEADER) String tokenValue,
+                                                           @RequestBody @Valid ScheduleSaveRequest request) {
+        Long userId = tokenProvider.getUserIdFromToken(tokenValue);
+        ScheduleResponse response = scheduleService.create(userId, request);
         return new ResponseEntity<>(response, HttpStatus.CREATED);
     }
 
     @GetMapping
-    public ResponseEntity<List<ScheduleResponse>> findAllSchedule(
-            @RequestParam(value = "page", defaultValue = "0") int page,
-            @RequestParam(value = "size", defaultValue = "10") int size) {
+    public ResponseEntity<List<ScheduleResponse>> findAllSchedule(@RequestParam(value = "page", defaultValue = "0") int page,
+                                                                  @RequestParam(value = "size", defaultValue = "10") int size) {
         Pageable pageable = PageRequest.of(page, size, Sort.by("modifiedAt").descending());
-        List<ScheduleResponse> responses = scheduleService.findAllSchedules(pageable);
+        List<ScheduleResponse> responses = scheduleService.findAllByPage(pageable);
         return new ResponseEntity<>(responses, HttpStatus.OK);
     }
 
@@ -47,21 +51,33 @@ public class ScheduleController {
 
     @PatchMapping("/{scheduleId}")
     public ResponseEntity<String> updateSchedule(@PathVariable("scheduleId") Long id,
+                                                 @CookieValue(value = AUTHORIZATION_HEADER) String tokenValue,
                                                  @RequestBody @Valid ScheduleUpdateRequest request) {
+        if (!tokenProvider.isUserAuthorizedWithToken(tokenValue)) {
+            throw new InvalidCredentialsException("해당 유저는 권한이 없습니다.");
+        }
         scheduleService.update(id, request);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PatchMapping("/assign/{scheduleId}")
     public ResponseEntity<String> assignUsersToSchedule(@PathVariable("scheduleId") Long id,
+                                                        @CookieValue(value = AUTHORIZATION_HEADER) String tokenValue,
                                                         @RequestBody AssignUsersRequest request) {
-        // 유저 id도 입력받기.
-        scheduleService.assignUsersToSchedule(id, request.userIds());
+        if (!tokenProvider.isUserAuthorizedWithToken(tokenValue)) {
+            throw new InvalidCredentialsException("해당 유저는 권한이 없습니다.");
+        }
+        Long userId = tokenProvider.getUserIdFromToken(tokenValue);
+        scheduleService.assignUsersToSchedule(id, userId, request.userIds());
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/{scheduleId}")
-    public ResponseEntity<String> deleteSchedule(@PathVariable("scheduleId") Long id) {
+    public ResponseEntity<String> deleteSchedule(@PathVariable("scheduleId") Long id,
+                                                 @CookieValue(value = AUTHORIZATION_HEADER) String tokenValue) {
+        if (!tokenProvider.isUserAuthorizedWithToken(tokenValue)) {
+            throw new InvalidCredentialsException("해당 유저는 권한이 없습니다.");
+        }
         scheduleService.delete(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
